# [HttpFoundation] Clear IpUtils cache to prevent memory leaks