| Q             | A
| ------------- | ---
| Branch?       | 6.3
| Bug fix?      | no
| New feature?  | no
| Deprecations? | no
| Tickets       | Fix #50032 
| License       | MIT
| Doc PR        | 

Fixes #50032 

IpUtils cache now uses LRU principal