@nicolas-grekas 2023-04-20
Can you run some bench comparing with/without LRU, also maybe without cache at all?