@nicolas-grekas 2023-04-20
> An option to improve a tiny bit would be to only cache IPv6 results since IPv4 checking seems to be a little bit faster than ANY caching.

let's do this yes! (with an inline comment reminding us about this)
let's drop the -v6 suffix while doing so