@nicolas-grekas 2023-04-20
> With LRU: 3.778 seconds, 2.097152 mb
> Without any Cache (IPv6): 5.741 seconds, 2.097152 mb

The difference is not that big actually, but let's still cache this :shrug: