@nicolas-grekas 2023-04-20
The cache has been introduced in https://github.com/symfony/symfony/pull/23098
What about dropping it altogether? The diff doesn't look much worth the added complexity to me.