@stof 2023-04-20
the benchmark being done seems to be testing the performance in case of cache misses though.

the cache was added to improve performance for the common case of checking the trusted ips many times, which will do cache hits for most checks.